(refactor) tool classes
