+++ src/tools/_delete_note.ts
@@ -1,5 +1,17 @@
+import { db } from "@/lib/db";
 import { ClaudeWeaverTool } from "./_tool"
 
-export default class DeleteNoteTool extends ClaudeWeaverTool {
+interface DeleteNoteToolParams {
+    id: string
+}
 
+export default class DeleteNoteTool extends ClaudeWeaverTool<DeleteNoteToolParams, boolean> {
+    async execute({id}: DeleteNoteToolParams){
+        const weaviateClient = await db.connect();
+        const notesCollection = weaviateClient.collections.get('Note');
+
+        await notesCollection.data.deleteById(id);
+
+        return true;
+    }
 }
