--- src/tools/_generate_web_with_google.ts
@@ -1,5 +1,0 @@
-import { ClaudeWeaverTool } from "./_tool"
-
-export default class SearchWebWithGoogleTool extends ClaudeWeaverTool {
-
-}
