+++ src/tools/_get_html_for_url.ts
@@ -1,5 +1,12 @@
 import { ClaudeWeaverTool } from "./_tool"
 
-export default class HTMLForURL extends ClaudeWeaverTool {
+interface HTMLForURLParams {
+    url: string
+}
 
+export default class HTMLForURL extends ClaudeWeaverTool<HTMLForURLParams, string> {
+    async execute({url}: HTMLForURLParams): Promise<string> {
+        const response = await fetch(url);
+        return await response.text(); 
+    }
 }
