+++ src/tools/_get_markdown_for_url.ts
@@ -1,5 +1,12 @@
 import { ClaudeWeaverTool } from "./_tool"
 
-export default class MDForURL extends ClaudeWeaverTool {
-
+interface MDForURLParams {
+    url: string
+}
+export default class MDForURL extends ClaudeWeaverTool<MDForURLParams, string> {
+    public async execute({url}: MDForURLParams): Promise<string> {
+        const jinaReaderUrl = `https://r.jina.ai/${url}`;
+        const response = await fetch(jinaReaderUrl);
+        return await response.text(); 
+    }
 }
